requirements.txt
django>4.2,<4.3
python-decouple
pillow
django-taggit